- Adição de arquivos separados para os types Card.ts e User.ts - Ajuste no uso desses tipos na store auth.ts e type Trade.ts
