+++ src/types/Card.ts
@@ -1,0 +1,7 @@
+export interface Card {
+  id: string
+  name: string
+  description: string
+  imageUrl: string
+  createdAt: string
+}
