+++ src/types/Trade.ts
@@ -1,3 +1,6 @@
+import type { Card } from './Card'
+import type { User } from './User'
+
 export interface Trade {
   id: string
   userId: string
@@ -6,22 +9,10 @@ export interface Trade {
   tradeCards: TradeCard[]
 }
 
-export interface User {
-  name: string
-}
-
 export interface TradeCard {
   id: string
   cardId: string
   tradeId: string
   type: 'OFFERING' | 'RECEIVING'
   card: Card
-}
-
-export interface Card {
-  id: string
-  name: string
-  description: string
-  imageUrl: string
-  createdAt: string
 }
