+++ src/types/User.ts
@@ -1,0 +1,9 @@
+import type { Card } from './Card'
+
+export interface User {
+  id: string
+  name: string
+  email: string
+  password: string
+  cards: Card[]
+}
